+++ framework/db/mssql/Schema.php
@@ -180,12 +180,7 @@ FROM [INFORMATION_SCHEMA].[TABLES] AS [t]
 WHERE [t].[table_schema] = :schema AND [t].[table_type] IN ('BASE TABLE', 'VIEW')
 ORDER BY [t].[table_name]
 SQL;
-        $tables = $this->db->createCommand($sql, [':schema' => $schema])->queryColumn();
-        $tables = array_map(static function ($item) {
-            return '[' . $item . ']';
-        }, $tables);
-
-        return $tables;
+        return $this->db->createCommand($sql, [':schema' => $schema])->queryColumn();
     }
 
     /**
@@ -202,6 +197,29 @@ SQL;
         }
 
         return null;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getSchemaMetadata($schema, $type, $refresh)
+    {
+        $metadata = [];
+        $methodName = 'getTable' . ucfirst($type);
+        $tableNames = array_map(function ($table) {
+            return $this->quoteSimpleTableName($table);
+        }, $this->getTableNames($schema, $refresh));
+        foreach ($tableNames as $name) {
+            if ($schema !== '') {
+                $name = $schema . '.' . $name;
+            }
+            $tableMetadata = $this->$methodName($name, $refresh);
+            if ($tableMetadata !== null) {
+                $metadata[] = $tableMetadata;
+            }
+        }
+
+        return $metadata;
     }
 
     /**
@@ -597,12 +615,7 @@ WHERE [t].[table_schema] = :schema AND [t].[table_type] = 'VIEW'
 ORDER BY [t].[table_name]
 SQL;
 
-        $views = $this->db->createCommand($sql, [':schema' => $schema])->queryColumn();
-        $views = array_map(static function ($item) {
-            return '[' . $item . ']';
-        }, $views);
-
-        return $views;
+        return $this->db->createCommand($sql, [':schema' => $schema])->queryColumn();
     }
 
     /**